requirements.txt
# SAM2 Video Training Dependencies
# Clean, minimal dependencies for the Lightning-based training system

# Core ML Framework
lightning>=2.0.0
torch>=2.0.0
torchvision>=0.15.0
torchmetrics>=1.0.0

# Configuration Management
hydra-core>=1.3.0
omegaconf>=2.3.0

# Data Processing
numpy>=1.24.0
Pillow>=9.0.0
pycocotools>=2.0.6
imageio>=2.28.0

# Logging
loguru>=0.7.0

# Experiment tracking
wandb>=0.15.0

# Utilities
tqdm>=4.64.0

# SAM2 Dependencies (install from official repository)
# git+https://github.com/facebookresearch/segment-anything-2.git

# Development Dependencies (optional)
# pytest>=7.0.0
# black>=22.0.0
# isort>=5.12.0
